pyproject.toml
[project]
name = "canadaca-chat"
version = "0.1.0"
description = "Bilingual RAG chat agent for Canada.ca"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "eval_type_backport>=0.2.0",  # Python 3.9 compatibility
    
    # LangChain & LangGraph
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.0.3",
    "langchain-aws>=0.2.0",
    "langchain-community>=0.0.20",
    "langsmith>=0.0.80",
    
    # AWS
    "boto3>=1.34.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.4",
    
    # Redis
    "redis>=5.0.0",
    
    # HTTP Client & Web Scraping
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    
    # Security
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "langdetect>=1.0.9",
    
    # Azure (optional, for Azure-specific services)
    "azure-identity>=1.15.0",
    "azure-ai-contentsafety>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",  # For testing FastAPI
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
